refactor: 순환 참조되는 훅을 하나로 통합
