+++ src/components/organisms/BalanceGameCreation/BalanceGameCreation.tsx
@@ -5,7 +5,6 @@ import DraftPostButton from '@/components/atoms/DraftPostButton/DraftPostButton'
 import { BalanceGameSet } from '@/types/game';
 import { useBalanceGameCreation } from '@/hooks/game/useBalanceGameCreation';
 import GameNavigationSection from '@/components/molecules/GameNavigationSection/GameNavigationSection';
-import { useStageNavigation } from '@/hooks/game/useStageNavigation';
 import useToastModal from '@/hooks/modal/useToastModal';
 import ToastModal from '@/components/atoms/ToastModal/ToastModal';
 import * as S from './BalanceGameCreation.style';
@@ -34,44 +33,22 @@ const BalanceGameCreation = ({
   const totalStage = 10;
   const { isVisible, modalText, showToastModal } = useToastModal();
 
-  const validateStage = (): true | string => {
-    if (!currentOptions || currentOptions.length < 2) {
-      return '옵션 데이터가 올바르지 않습니다!';
-    }
-
-    if (!currentOptions[0]?.name.trim() || !currentOptions[1]?.name.trim()) {
-      return '모든 옵션의 설명을 입력해주세요!';
-    }
-
-    const hasBothImages =
-      currentOptions[0]?.imgUrl.trim() && currentOptions[1]?.imgUrl.trim();
-    const hasNoImages =
-      !currentOptions[0]?.imgUrl.trim() && !currentOptions[1]?.imgUrl.trim();
-
-    if (!(hasBothImages || hasNoImages)) {
-      return 'A와 B의 이미지가 모두 없거나 모두 있어야 합니다!';
-    }
-
-    return true;
-  };
-
-  const { currentStage, clearInput, handleNextStage, handlePrevStage } =
-    useStageNavigation(totalStage, validateStage, showToastModal);
-
   const {
     games,
+    currentStage,
     currentOptions,
     currentDescription,
+    clearInput,
+    handleNextStage,
+    handlePrevStage,
+    handleStageDescriptionChange,
     handleOptionChange,
     handleDescriptionChange,
-    handleStageDescriptionChange,
-  } = useBalanceGameCreation(currentStage, loadedGames, totalStage);
+  } = useBalanceGameCreation(showToastModal, totalStage, loadedGames);
 
   useEffect(() => {
-    if (currentOptions.length > 0) {
-      onGamesUpdate(games);
-    }
-  }, [currentOptions, games, onGamesUpdate]);
+    onGamesUpdate(games);
+  }, [games, onGamesUpdate]);
 
   return (
     <div css={S.pageContainer}>
