+++ src/hooks/game/useBalanceGameCreation.ts
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState, useRef } from 'react';
 import { BalanceGameOption, BalanceGameSet } from '@/types/game';
 import {
   createInitialGameStages,
@@ -6,15 +6,19 @@ import {
 } from '@/utils/balanceGameUtils';
 
 export const useBalanceGameCreation = (
-  currentStage: number,
-  loadedGames?: BalanceGameSet[],
+  showToast: (message: string) => void,
   totalStage: number = 10,
+  loadedGames?: BalanceGameSet[],
 ) => {
   const [games, setGames] = useState<BalanceGameSet[]>(
     loadedGames || createInitialGameStages(totalStage),
   );
+  const [currentStage, setCurrentStage] = useState(0);
   const [currentOptions, setCurrentOptions] = useState<BalanceGameOption[]>([]);
   const [currentDescription, setCurrentDescription] = useState<string>('');
+  const [clearInput, setClearInput] = useState(false);
+
+  const timerRef = useRef<number | NodeJS.Timeout | null>(null);
 
   useEffect(() => {
     if (loadedGames) {
@@ -27,6 +31,24 @@ export const useBalanceGameCreation = (
     setCurrentOptions(stage.gameOptions);
     setCurrentDescription(stage.description);
   }, [currentStage, games]);
+
+  useEffect(() => {
+    if (clearInput) {
+      if (timerRef.current !== null) {
+        clearTimeout(timerRef.current);
+      }
+
+      timerRef.current = setTimeout(() => {
+        setClearInput(false);
+      }, 500);
+    }
+
+    return () => {
+      if (timerRef.current !== null) {
+        clearTimeout(timerRef.current);
+      }
+    };
+  }, [clearInput]);
 
   const updateOption = (
     stageIndex: number,
@@ -55,6 +77,52 @@ export const useBalanceGameCreation = (
     );
   };
 
+  const validateStage = (): true | string => {
+    if (!currentOptions[0]?.name.trim() || !currentOptions[1]?.name.trim()) {
+      return '모든 옵션의 설명을 입력해주세요!';
+    }
+
+    const hasBothImages =
+      currentOptions[0]?.imgUrl.trim() && currentOptions[1]?.imgUrl.trim();
+    const hasNoImages =
+      !currentOptions[0]?.imgUrl.trim() && !currentOptions[1]?.imgUrl.trim();
+
+    if (!(hasBothImages || hasNoImages)) {
+      return 'A와 B의 이미지가 모두 없거나 모두 있어야 합니다!';
+    }
+
+    return true;
+  };
+
+  const handleNextStage = () => {
+    const validationResult = validateStage();
+    if (currentStage < totalStage - 1) {
+      if (validationResult === true) {
+        setClearInput(true);
+        setCurrentStage((prev) => prev + 1);
+      } else {
+        showToast(validationResult);
+      }
+    }
+  };
+
+  const handlePrevStage = () => {
+    if (currentStage > 0) {
+      setClearInput(true);
+      setCurrentStage((prev) => prev - 1);
+    }
+  };
+
+  const handleStageDescriptionChange = (newDescription: string) => {
+    setCurrentDescription(newDescription);
+
+    setGames((prevGames) =>
+      prevGames.map((game, idx) =>
+        idx === currentStage ? { ...game, description: newDescription } : game,
+      ),
+    );
+  };
+
   const handleOptionChange =
     (optionType: 'A' | 'B') => (event: React.ChangeEvent<HTMLInputElement>) => {
       updateOption(currentStage, optionType, { name: event.target.value });
@@ -68,22 +136,17 @@ export const useBalanceGameCreation = (
     updateOption(currentStage, optionType, { description: value });
   };
 
-  const handleStageDescriptionChange = (newDescription: string) => {
-    setCurrentDescription(newDescription);
-
-    setGames((prevGames) =>
-      prevGames.map((game, idx) =>
-        idx === currentStage ? { ...game, description: newDescription } : game,
-      ),
-    );
-  };
-
   return {
     games,
+    currentStage,
     currentOptions,
     currentDescription,
+    clearInput,
+    handleNextStage,
+    handlePrevStage,
+    handleStageDescriptionChange,
     handleOptionChange,
     handleDescriptionChange,
-    handleStageDescriptionChange,
+    validateStage,
   };
 };
